Improve API url and key handling
API url and keys are stored in environment variables. This is a
recommended method of storing such information.

The Github runners set theese variables using secrets.

For the user it is more annoying. Currently the url and key must be set
every time, because the variables are note persistent. One way of
solving this might be to follow this guide:
https://pybit.es/persistent-environment-variables.html